+++ identity/idp_group_mapping.go
@@ -47,7 +47,7 @@ type IdpGroupMapping struct {
 	LifecycleState IdpGroupMappingLifecycleStateEnum `mandatory:"true" json:"lifecycleState,omitempty"`
 
 	// The detailed status of INACTIVE lifecycleState.
-	InactiveStatus *int64 `mandatory:"false" json:"inactiveStatus,omitempty"`
+	InactiveStatus *int `mandatory:"false" json:"inactiveStatus,omitempty"`
 }
 
 func (model IdpGroupMapping) String() string {
